+++ packages/ui/src/components/button/button.types.ts
@@ -55,6 +55,11 @@ export type ButtonProps = {
    */
   look?: Variants['look'];
   /**
+   * Removes horizontal padding from the 'link' look button
+   * @default false
+   */
+  removeLinkPadding?: boolean;
+  /**
    * Size of the button
    * @default medium
    */