+++ packages/ui/src/components/modal/components/modal-dialog/components/modal-dialog-body/modal-dialog-body.component.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+'use client';
 
 import { useModalDialogContext } from '../../modal-dialog.component.js';
 
@@ -6,10 +6,12 @@ import { styles as modalBodyStyles } from './modal-dialog-body.styles.js';
 import { type ModalDialogBodyProps } from './modal-dialog-body.types.js';
 
 export function ModalDialogBody({ className, children, ...props }: ModalDialogBodyProps) {
-  const { size } = useModalDialogContext();
-  const styles = modalBodyStyles({ size });
+  const { size, scrollingRef, canScroll, compact, footerPresent, scrollAtBottom } = useModalDialogContext();
+
+  const styles = modalBodyStyles({ size, canScroll, scrollAtBottom, compact, footerPresent });
+
   return (
-    <div className={styles.base({ className })} {...props}>
+    <div className={styles.base({ className })} ref={scrollingRef} {...props}>
       {children}
     </div>
   );