+++ packages/ui/src/components/popover/components/panel/panel.styles.ts
@@ -10,7 +10,7 @@ export const styles = tv(
         after:left-[1.5px] after:top-0 after:size-0 after:border-x-[6.5px] after:border-t-[11px] after:border-x-[transparent] after:border-t-white
       `,
       closeBtn: 'absolute right-1 top-1 h-3 p-0 hover:opacity-80',
-      content: 'w-[17.625rem] rounded-[3px] bg-white py-4 pl-3 pr-5',
+      content: 'w-[18.75rem] rounded-[3px] bg-white py-4 pl-3 pr-5',
       heading: 'typography-body-9 mb-2 font-medium text-text focus-visible:focus-outline',
       body: 'typography-body-10 text-text focus-visible:focus-outline',
     },